Return more icons from _NET_WM_ICON
EWMH says that _NET_WM_ICON is an array of icon. That means that it can contain
more than one icon. This way, apps can provide icons of different sizes. Common
sizes seem to be 16x16, 32x32 and 48x48.

This commit makes it possible to access all the icons in _NET_WM_ICON via
iterators instead of just the first one.

Thanks to Arnaud Fontaine for his help with the API.

Signed-off-by: Uli Schlachter <psychon@znc.in>
Signed-off-by: Arnaud Fontaine <arnau@debian.org>