xcb_ewmh_get_wm_icon_reply: Fix crash on error
When the GetProperty request fails due to an error,
xcb_ewmh_get_wm_icon_from_reply will be called with a NULL pointer for the
reply. This function would then call xcb_get_property_value_length on this NULL
pointer which caused a crash.

Fix this by moving the NULL-pointer check before the call to value_length().

Signed-off-by: Uli Schlachter <psychon@znc.in>
Reviewed-by: Arnaud Fontaine <arnau@debian.org>